+++ service_restaurant/src/main/kotlin/org/thechance/service_restaurant/api/endpoints/CategoryRoutes.kt
@@ -7,10 +7,11 @@ import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import org.koin.ktor.ext.inject
 import org.thechance.service_restaurant.api.models.CategoryDto
+import org.thechance.service_restaurant.api.models.mappers.toDto
+import org.thechance.service_restaurant.api.models.mappers.toEntity
 import org.thechance.service_restaurant.usecase.category.CategoryUseCasesContainer
 import org.thechance.service_restaurant.api.utils.extractInt
 import org.thechance.service_restaurant.api.utils.extractString
-import org.thechance.service_restaurant.entity.toDto
 
 
 fun Route.categoryRoutes() {
