+++ service_restaurant/src/main/kotlin/org/thechance/service_restaurant/data/collection/CategoryCollection.kt
@@ -5,8 +5,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import org.bson.codecs.pojo.annotations.BsonId
 import org.bson.types.ObjectId
-import org.litote.kmongo.Id
-import org.thechance.service_restaurant.entity.Category
 
 @Serializable
 @SerialName("categoryCollection")
@@ -19,14 +17,5 @@ data class CategoryCollection(
     @SerialName("_id")
     val id = ObjectId()
     val isDeleted: Boolean = false
-
-    fun toEntity(): Category {
-        return Category(
-            id = id.toString(),
-            name = name,
-            isDeleted = isDeleted,
-        )
-    }
 }
 
-fun List<CategoryCollection>.toEntity(): List<Category> = map { it.toEntity() }
