+++ service_restaurant/src/main/kotlin/org/thechance/service_restaurant/entity/Restaurant.kt
@@ -16,32 +16,5 @@ data class Restaurant(
     val isDeleted: Boolean = false
 //    val address: ObjectId,
 //    val meals: List<ObjectId>,
-) {
+)
 
-    fun toDto(): RestaurantDto {
-        return RestaurantDto(
-            id = id ?: "",
-            name = name,
-            description = description,
-            priceLevel = priceLevel,
-            rate = rate,
-            phone = phone,
-            openingTime = openingTime,
-            closingTime = closingTime
-        )
-    }
-
-    fun toCollection(): RestaurantCollection {
-        return RestaurantCollection(
-            name = name,
-            description = description,
-            priceLevel = priceLevel,
-            rate = rate,
-            phone = phone,
-            openingTime = openingTime,
-            closingTime = closingTime,
-        )
-    }
-}
-
-fun List<Restaurant>.toDto(): List<RestaurantDto> = map { it.toDto() }
